Update react-resize-detector, react-cookie, react-calendar
