+++ package.json
@@ -130,9 +130,9 @@
     "jsdom": "^26.0.0",
     "lodash": "^4.17.21",
     "react": "^18.3.1",
-    "react-calendar": "^4.0.0",
+    "react-calendar": "^5.1.0",
     "react-chat-widget-react-18": "^1.0.0",
-    "react-cookie": "^4.1.1",
+    "react-cookie": "^8.0.1",
     "react-copy-to-clipboard": "^5.0.3",
     "react-dom": "^18.3.1",
     "react-i18next": "^15.5.1",
@@ -141,7 +141,7 @@
     "react-intersection-observer": "^9.1.0",
     "react-markdown": "^10.1.0",
     "react-mentions": "^4.4.7",
-    "react-resize-detector": "^7.0.0",
+    "react-resize-detector": "^12.0.2",
     "react-router-dom": "^7.4.0",
     "react-scroll": "^1.8.7",
     "rehype-parse": "^9.0.1",
@@ -170,8 +170,6 @@
     "@graphql-codegen/typescript-react-apollo": "^4.3.2",
     "@sentry/vite-plugin": "^2.7.1",
     "@testing-library/jest-dom": "^6.6.3",
-    "@types/js-cookie": "^3.0.2",
-    "@types/react-calendar": "^3.9.0",
     "@types/react-copy-to-clipboard": "^5.0.0",
     "@types/react-mentions": "^4.1.8",
     "@typescript-eslint/parser": "^7.16.1",
